Module type Swresample.AudioData

A data type a converter reads or writes. The modules of this type are the ones at the end of this interface, from Bytes to DblPlanarFrame; a program passes two of them to Make.

type t

val kind : t kind