Module Swresample

This module perform audio resampling, rematrixing and sample format conversion operations.

val version : Avutil.version

module type AudioData = sig ... end

Audio data modules for Swresample module input and output parameterization.

type options = [

| Swresample__.Swresample_options.dither_type

| Swresample__.Swresample_options.engine

| Swresample__.Swresample_options.filter_type

]

type ('i, 'o) ctx

module Make (I : AudioData) (O : AudioData) : sig ... end

Functor building an implementation of the swresample structure with parameterized input an output audio data types

module Bytes : sig ... end

Byte string with undefined sample format for interleaved channels. The sample format must be passed to the create function.

module U8Bytes : sig ... end

Unsigned 8 bit sample format byte string for interleaved channels.

module S16Bytes : sig ... end

Signed 16 bit sample format byte string for interleaved channels.

module S32Bytes : sig ... end

Signed 32 bit sample format byte string for interleaved channels.

module FltBytes : sig ... end

Float 32 bit sample format byte string for interleaved channels.

module DblBytes : sig ... end

Float 64 bit sample format byte string for interleaved channels.

module U8PlanarBytes : sig ... end

Unsigned 8 bit sample format byte string for planar channels.

module S16PlanarBytes : sig ... end

Signed 16 bit sample format byte string for planar channels.

module S32PlanarBytes : sig ... end

Signed 32 bit sample format byte string for planar channels.

module FltPlanarBytes : sig ... end

Float 32 bit sample format byte string for planar channels.

module DblPlanarBytes : sig ... end

Float 64 bit sample format byte string for planar channels.

module FloatArray : sig ... end

Float 64 bit sample format array for interleaved channels.

module PlanarFloatArray : sig ... end

Float 64 bit sample format array for planar channels.

type u8ba = (int, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

type s16ba = (int, Stdlib.Bigarray.int16_signed_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

type s32ba = (int32, Stdlib.Bigarray.int32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

type f32ba = (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

type f64ba = (float, Stdlib.Bigarray.float64_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

module U8BigArray : sig ... end

Unsigned 8 bit sample format bigarray for interleaved channels.

module S16BigArray : sig ... end

Signed 16 bit sample format bigarray for interleaved channels.

module S32BigArray : sig ... end

Signed 32 bit sample format bigarray for interleaved channels.

module FltBigArray : sig ... end

Float 32 bit sample format bigarray for interleaved channels.

module DblBigArray : sig ... end

Float 64 bit sample format bigarray for interleaved channels.

module U8PlanarBigArray : sig ... end

Unsigned 8 bit sample format bigarray for planar channels.

module S16PlanarBigArray : sig ... end

Signed 16 bit sample format bigarray for planar channels.

module S32PlanarBigArray : sig ... end

Signed 32 bit sample format bigarray for planar channels.

module FltPlanarBigArray : sig ... end

Float 32 bit sample format bigarray for planar channels.

module DblPlanarBigArray : sig ... end

Float 64 bit sample format bigarray for planar channels.

module Frame : sig ... end

Audio frame with undefined sample format. The sample format must be passed to the create function.

module U8Frame : sig ... end

Unsigned 8 bit sample format audio frame for interleaved channels.

module S16Frame : sig ... end

Signed 16 bit sample format audio frame for interleaved channels.

module S32Frame : sig ... end

Signed 32 bit sample format audio frame for interleaved channels.

module FltFrame : sig ... end

Float 32 bit sample format audio frame for interleaved channels.

module DblFrame : sig ... end

Float 64 bit sample format audio frame for interleaved channels.

module U8PlanarFrame : sig ... end

Unsigned 8 bit sample format audio frame for planar channels.

module S16PlanarFrame : sig ... end

Signed 16 bit sample format audio frame for planar channels.

module S32PlanarFrame : sig ... end

Signed 32 bit sample format audio frame for planar channels.

module FltPlanarFrame : sig ... end

Float 32 bit sample format audio frame for planar channels.

module DblPlanarFrame : sig ... end

Float 64 bit sample format audio frame for planar channels.