Module Swresample.Bytes

Byte string with undefined sample format for interleaved channels. The sample format must be passed to the create function.

type t = bytes

val vk : {vector_kind}1

val sf : Avutil.Sample_format.t